Fixed issues with version 1.13.2
