+++ src/main/java/game/data/registries/RegistryLoader.java
@@ -60,7 +60,11 @@ public class RegistryLoader {
      * server jar couldn't yet generate reports at this point.
      */
     private boolean hasExistingReports() {
-        return version.equals("1.12.2") || blocksPath.toFile().exists() && registryPath.toFile().exists();
+        if (version.equals("1.13.2")) {
+            System.out.println("No item registry for 1.13.2. We won't be able to save chests contents :(");
+        }
+
+        return version.equals("1.12.2") || blocksPath.toFile().exists();
     }
 
     /**
